fixing query runner round trip OOM issue
issue occur when completion is not properly called or waited during shutdown
the resolution is that only the receiving end is able to make onComplete
call towards the sending end. this is because when the sending end calls
onComplete it doesn't mean the receiving end actually finish processing
all the data (thus hasn't finish responding with buffer size return
status messeages).

Thus
1. sending side indicate the data is complete via sending a metadata
   only block
2. receiving side will response with all the buffer size status message,
   then calls onComplete.
3. sending side waits for call onComplete or waits for termination via
   timeout.
