+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/runtime/mailbox/channel/GrpcMailboxServer.java
@@ -4,6 +4,7 @@ import io.grpc.Server;
 import io.grpc.ServerBuilder;
 import io.grpc.stub.StreamObserver;
 import java.io.IOException;
+import java.util.concurrent.TimeUnit;
 import org.apache.pinot.common.proto.Mailbox;
 import org.apache.pinot.common.proto.PinotMailboxGrpc;
 import org.apache.pinot.query.runtime.mailbox.GrpcMailboxService;
@@ -14,6 +15,7 @@ import org.slf4j.LoggerFactory;
 
 public class GrpcMailboxServer extends PinotMailboxGrpc.PinotMailboxImplBase {
   private static final Logger LOGGER = LoggerFactory.getLogger(GrpcMailboxServer.class);
+  private static final long DEFAULT_GRPC_MAILBOX_SERVER_TIMEOUT = 10000L;
 
   private final GrpcMailboxService _mailboxService;
   private final Server _server;
@@ -34,7 +36,7 @@ public class GrpcMailboxServer extends PinotMailboxGrpc.PinotMailboxImplBase {
 
   public void shutdown() {
     try {
-      _server.shutdown().awaitTermination();
+      _server.shutdown().awaitTermination(DEFAULT_GRPC_MAILBOX_SERVER_TIMEOUT, TimeUnit.MILLISECONDS);
     } catch (InterruptedException e) {
       throw new RuntimeException(e);
     }
